Avoid just using namespace as it is there in multiple examples
https://github.com/aws-samples/cdk-eks-karpenter/issues/106

It resolves this issue.

Avoid just using namespace as logical id in addManifest method as it is used in multiple examples where namespaces are created. 

Users who build their clusters from ground up using this addManifest might have many resources in their namespace which they might not be able to delete.

Many examples have this code -
const namespace = this.cluster.addManifest('namespace', {
      apiVersion: 'v1',
      kind: 'Namespace',
      metadata: {
        name: this.namespace,
      },
    });

Due to this the logical id will be common for many other operators which will lead to the error pointed out in the issue.

This will be a breaking change for many users upgrading to this code.